# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/huespedes": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener lista de huéspedes paginada con filtros dinámicos",
                "operationId": "3ad04ab3a6eb86d1d29da81aa09340a7",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre del huésped para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellido",
                        "in": "query",
                        "description": "Apellido del huésped para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dniPasaporte",
                        "in": "query",
                        "description": "DNI o Pasaporte del huésped para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fechaCheckin",
                        "in": "query",
                        "description": "Fecha de check-in para filtrar",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fechaCheckout",
                        "in": "query",
                        "description": "Fecha de check-out para filtrar",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "habitacion_id",
                        "in": "query",
                        "description": "ID de la habitación asociada para filtrar",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes filtrada y paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Crear un nuevo huésped",
                "operationId": "ccc51562d959dbc9cede5e9dcc029082",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHuespedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Huésped creado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/huespedes/{idHuesped}": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener un huésped por ID",
                "operationId": "de14a0c968280d9ea3ab7687648b3365",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del huésped"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Actualizar un huésped",
                "operationId": "f6a7a88a4b814d69af67d386f2f36ccb",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHuespedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Huésped actualizado correctamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Eliminar un huésped",
                "operationId": "79432f27cac65a9296d9a1545d55b3db",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped eliminado"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Actualizar un huésped",
                "operationId": "3e3fdc3a3b4bd24d3ec699db7608ac18",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHuespedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Huésped actualizado correctamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
        "/api/huespedes/{idHuesped}/habitacion": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener la habitación de un huésped",
                "operationId": "96555c3c649f4021b64e7c1665dc483c",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación del huésped"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
        "/api/huespedes/{idHuesped}/reservar/{idHabitacion}": {
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Reservar una habitación para un huésped",
                "operationId": "2f33094ba31c05d8468828b0d506d1ab",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idHabitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Habitación reservada correctamente"
                    },
                    "400": {
                        "description": "Habitación no disponible"
                    },
                    "404": {
                        "description": "Huésped o habitación no encontrado"
                    }
                }
            }
        },
        "/api/huespedes/{idHuesped}/checkout": {
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Registrar el check-out del huésped y liberar la habitación",
                "operationId": "794a8f940674b5f768a3864c31720ef8",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check-out registrado correctamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreHuespedRequest": {
                "title": "Store Huesped Request",
                "description": "Datos requeridos para crear un nuevo huésped",
                "required": [
                    "nombre",
                    "apellido",
                    "dniPasaporte"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del huésped",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del huésped",
                        "type": "string"
                    },
                    "dniPasaporte": {
                        "description": "DNI o pasaporte del huésped",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PutHuespedRequest": {
                "title": "Put Huesped Request",
                "description": "Datos requeridos para actualizar un huésped",
                "required": [
                    "nombre",
                    "apellido",
                    "dniPasaporte"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del huésped",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del huésped",
                        "type": "string"
                    },
                    "dniPasaporte": {
                        "description": "DNI o pasaporte del huésped",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}